Reduce dupplication with static methods
